+++ biogassim/Properties/__init__.py
@@ -1,6 +1,13 @@
 """Pacote Properties: componentes puros e misturas."""
 from .Amines import amine_cp, amine_density, amine_viscosity, heat_of_absorption
 from .components import Component, ShomateCp, all_components, get
+from .GasProperties import (
+    GasProperties,
+    compressibility,
+    mixture_molar_mass,
+    mixture_properties,
+    normalize_composition,
+)
 from .Mixtures import cp_ideal_mixture, molar_weight, wilke_viscosity
 from .Water import water_cp, water_density, water_surface_tension, water_viscosity
 
@@ -9,4 +16,6 @@ __all__ = [
     "molar_weight", "cp_ideal_mixture", "wilke_viscosity",
     "water_density", "water_viscosity", "water_cp", "water_surface_tension",
     "amine_density", "amine_viscosity", "amine_cp", "heat_of_absorption",
+    "GasProperties", "mixture_properties", "normalize_composition",
+    "mixture_molar_mass", "compressibility",
 ]